--- src/types/btn/RentalLocationBtnTypes.ts
@@ -1,3 +1,0 @@
-export type TRentalLocationBtn = {
-  text: string[];
-};
