--- src/types/label/RentalLocationTitleTypes.ts
@@ -1,4 +1,0 @@
-export type TRentalLocationTitle = {
-  title: string;
-  category: string[];
-};
